pyproject.toml
[tool.poetry]
name = "homework4"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^1.8.1"
uvicorn = "^0.13.4"
python-multipart = "^0.0.5"
aioredis = "^1.3.1"
aiosqlite = "^0.17.0"
SQLAlchemy = "1.4.3"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-1.8.1%2Bcpu-cp39-cp39-linux_x86_64.whl"}
transformers = "^4.5.1"
sentence-transformers = "^1.1.0"
alembic = "^1.5.8"
pytest-asyncio = "^0.15.1"
pytest = "^6.2.3"
pytest-mock = "^3.6.0"
httpx = "^0.18.1"
fakeredis = "^1.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
autoflake = "^1.4"
isort = "^5.7.0"
black = "^20.8b1"
unify = "^0.5"
flake8 = "^3.8.4"
pylint = "^2.7.1"
mypy = "^0.812"
pytest-deadfixtures = "^2.2.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
